[Issue ordering] Using integer instead of float

+++ db/migrate/20170131221752_add_relative_position_to_issues.rb
@@ -24,7 +24,7 @@ class AddRelativePositionToIssues < ActiveRecord::Migration
   # disable_ddl_transaction!
 
   def change
-    add_column :issues, :relative_position, :float
+    add_column :issues, :relative_position, :integer
 
     add_index :issues, :relative_position
   end
